writeFileSafely関数をsrc/backend/electron以下に移動
