+++ src/helpers/fileHelper.ts
@@ -1,19 +1,4 @@
-import fs from "fs";
-import { moveFile } from "move-file";
 import { ResultError } from "@/type/result";
-
-export async function writeFileSafely(
-  path: string,
-  data: string | NodeJS.ArrayBufferView,
-) {
-  // ファイル書き込みに失敗したときに設定が消えないように、tempファイル書き込み後上書き移動する
-  const temp_path = `${path}.tmp`;
-  await fs.promises.writeFile(temp_path, data);
-
-  await moveFile(temp_path, path, {
-    overwrite: true,
-  });
-}
 
 /** ファイル書き込み時のエラーメッセージを生成する */
 // instanceof ResultErrorで生まれるResultError<any>を受け取れるようにするため、anyを許容する
